pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Dárley Raffael <darleyraffael@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = "^0.109.0"
uvicorn = "^0.26.0"
pydantic = {extras = ["email"], version = "^2.5.3"}
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
pydantic-settings = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"


[tool.poetry.group.dev.dependencies]
taskipy = "^1.12.2"
pytest-cov = "^4.1.0"
httpx = "^0.26.0"
pytest = "^7.4.4"
factory-boy = "^3.3.0"
freezegun = "^1.4.0"
ipykernel = "^6.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.taskipy.tasks]
dev = "uvicorn app.main:app --reload"
test = "pytest -s -x --cov=app -vv"
post_test = "coverage html"


[tool.pytest.ini_options]
pythonpath = "."